requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
google-api-python-client==1.7.11
google-auth==1.11.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gql==0.4.0
graphql-core==2.3.1
httplib2==0.17.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.4
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==1.6.1
six==1.14.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
yarl==1.4.2
